+++ andes/models/info.py
@@ -19,6 +19,9 @@ class Summary(ModelData, Model):
 
         self.field = DataParam(info='field name')
         self.comment = DataParam(info='information, comment, or anything')
+        self.comment2 = DataParam(info='comment field 2')
+        self.comment3 = DataParam(info='comment field 3')
+        self.comment4 = DataParam(info='comment field 4')
 
         Model.__init__(self, system, config)
         self.group = 'Information'